\name{mixed_model}
\alias{mixed_model}

\title{
Generalized Linear Mixed Effects Models
}

\description{
Fits generalized linear mixed effects models under maximum likelihood using adaptive 
Gaussian quadrature.
}

\usage{
mixed_model(fixed, random, data, family = NULL, na.action = na.exclude,
  n_phis = NULL, initial_values = NULL, control = list(), \dots)
}

\arguments{
\item{fixed}{.}
\item{random}{.}
\item{data}{.}
\item{family}{.}
\item{na.action}{.}
\item{n_phis}{.}
\item{initial_values}{.}
\item{control}{.}
\item{\dots}{.}
}

\details{
...
}

\value{
...
}

\author{
Dimitris Rizopoulos \email{d.rizopoulos@erasmusmc.nl}
}

\examples{
1 + 1
}