\name{fixef}
\alias{fixef}
\alias{fixef.MixMod}
\alias{ranef}
\alias{ranef.MixMod}

\title{
...
}

\description{
...
}

\usage{

fixef(object, \dots)

ranef(object, \dots)

\method{fixef}{MixMod}(object, \dots)

\method{ranef}{MixMod}(object, \dots)

}

\arguments{
\item{object}{.}
\item{\dots}{.}
}

\details{
...
}

\value{
...
}

\author{
Dimitris Rizopoulos \email{d.rizopoulos@erasmusmc.nl}
}

\examples{
1 + 1
}