\name{effectPlotData}
\alias{effectPlotData}
\alias{effectPlotData.MixMod}

\title{
...
}

\description{
...
}

\usage{

effectPlotData(object, newdata, level, \dots)

\method{effectPlotData}{MixMod}(object, newdata, 
   level = 0.95, marginal = FALSE, \dots)
}

\arguments{
\item{object}{.}
\item{newdata}{.}
\item{level}{.}
\item{marginal}{.}
\item{\dots}{.}
}

\details{
...
}

\value{
...
}

\author{
Dimitris Rizopoulos \email{d.rizopoulos@erasmusmc.nl}
}

\examples{
1 + 1
}