\name{marginal_coefs}
\alias{marginal_coefs}

\title{
...
}

\description{
...
}

\usage{
marginal_coefs(object, std_errors = FALSE, link_fun = NULL, 
   M = 3000, K = 100, seed = 1, 
   cores = max(parallel::detectCores() - 1, 1), \dots)
}

\arguments{
\item{object}{.}
\item{std_errors}{.}
\item{link_fun}{.}
\item{M}{.}
\item{K}{.}
\item{seed}{.}
\item{cores}{.}
\item{\dots}{.}
}

\details{
...
}

\value{
...
}

\author{
Dimitris Rizopoulos \email{d.rizopoulos@erasmusmc.nl}
}

\examples{
1 + 1
}